WINDUP-141
Unwrap EjbRule
